Move common data plugins up a level - Replace reducers dir with reducers.js - Update plugins reducer to use object instead of array
